Admin Edit Functionality for all entities except attendance

+++ src/main/java/com/grownited/repository/SubjectRepository.java
@@ -1,11 +1,15 @@
 package com.grownited.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.grownited.entity.SubjectEntity;
 
 @Repository
-public interface SubjectRepository extends JpaRepository<SubjectEntity, Integer>{
+public interface SubjectRepository extends JpaRepository<SubjectEntity, Long>{
+
+	
 
 }
